+++ muicebot/config.py
@@ -18,6 +18,8 @@ class PluginConfig(BaseModel):
     """日志等级"""
     telegram_proxy: str | None = None
     """telegram代理，这个配置项用于获取图片时使用"""
+    plugins_dir: list = ["./muicebot/plugins", "./muicebot/builtin_plugins"]
+    """自定义插件加载目录"""
     including_nicknames: list = ["沐雪", "muice", "雪雪", "Muice"]
     """句中触发词列表"""
     including_nicknames_trigger_coefficient: float = 0.05
